+++ resources/lib/ytmusicapi/mixins/_protocol.py
@@ -1,6 +1,6 @@
 """protocol that defines the functions available to mixins"""
 
-from typing import Dict, Optional, Protocol
+from typing import Optional, Protocol, Dict
 
 from requests import Response
 
@@ -20,10 +20,10 @@ class MixinProtocol(Protocol):
     def _check_auth(self) -> None:
         """checks if self has authentication"""
 
-    def _send_request(self, endpoint: str, body: Dict, additionalParams: str = "") -> Dict:
+    def _send_request(self, endpoint: str, body: dict, additionalParams: str = "") -> dict:
         """for sending post requests to YouTube Music"""
 
-    def _send_get_request(self, url: str, params: Optional[Dict] = None) -> Response:
+    def _send_get_request(self, url: str, params: Optional[dict] = None) -> Response:
         """for sending get requests to YouTube Music"""
 
     @property