+++ resources/lib/ytmusicapi/mixins/explore.py
@@ -5,7 +5,7 @@ from ytmusicapi.parsers.explore import *
 
 
 class ExploreMixin(MixinProtocol):
-    def get_mood_categories(self) -> Dict:
+    def get_mood_categories(self) -> dict:
         """
         Fetch "Moods & Genres" categories from YouTube Music.
 
@@ -62,7 +62,7 @@ class ExploreMixin(MixinProtocol):
 
         return sections
 
-    def get_mood_playlists(self, params: str) -> List[Dict]:
+    def get_mood_playlists(self, params: str) -> List[dict]:
         """
         Retrieve a list of playlists for a given "Moods & Genres" category.
 
@@ -88,7 +88,7 @@ class ExploreMixin(MixinProtocol):
 
         return playlists
 
-    def get_charts(self, country: str = "ZZ") -> Dict:
+    def get_charts(self, country: str = "ZZ") -> dict:
         """
         Get latest charts data from YouTube Music: Top songs, top videos, top artists and top trending videos.
         Global charts have no Trending section, US charts have an extra Genres section with some Genre charts.