+++ resources/lib/ytmusicapi/mixins/uploads.py
@@ -1,6 +1,5 @@
-import ntpath
-import os
-from typing import Dict, List, Optional, Union
+from pathlib import Path
+from typing import Optional, Union, List
 
 import requests
 
@@ -24,7 +23,7 @@ from ._utils import prepare_order_params, validate_order_parameter
 
 
 class UploadsMixin(MixinProtocol):
-    def get_library_upload_songs(self, limit: Optional[int] = 25, order: Optional[str] = None) -> List[Dict]:
+    def get_library_upload_songs(self, limit: Optional[int] = 25, order: Optional[str] = None) -> List[dict]:
         """
         Returns a list of uploaded songs
 
@@ -71,7 +70,7 @@ class UploadsMixin(MixinProtocol):
 
         return songs
 
-    def get_library_upload_albums(self, limit: Optional[int] = 25, order: Optional[str] = None) -> List[Dict]:
+    def get_library_upload_albums(self, limit: Optional[int] = 25, order: Optional[str] = None) -> List[dict]:
         """
         Gets the albums of uploaded songs in the user's library.
 
@@ -92,7 +91,7 @@ class UploadsMixin(MixinProtocol):
 
     def get_library_upload_artists(
         self, limit: Optional[int] = 25, order: Optional[str] = None
-    ) -> List[Dict]:
+    ) -> List[dict]:
         """
         Gets the artists of uploaded songs in the user's library.
 
@@ -111,7 +110,7 @@ class UploadsMixin(MixinProtocol):
             response, lambda additionalParams: self._send_request(endpoint, body, additionalParams), limit
         )
 
-    def get_library_upload_artist(self, browseId: str, limit: int = 25) -> List[Dict]:
+    def get_library_upload_artist(self, browseId: str, limit: int = 25) -> List[dict]:
         """
         Returns a list of uploaded tracks for the artist.
 
@@ -160,7 +159,7 @@ class UploadsMixin(MixinProtocol):
 
         return items
 
-    def get_library_upload_album(self, browseId: str) -> Dict:
+    def get_library_upload_album(self, browseId: str) -> dict:
         """
         Get information and tracks of an album associated with uploaded tracks
 
@@ -212,11 +211,12 @@ class UploadsMixin(MixinProtocol):
         self._check_auth()
         if not self.auth_type == AuthType.BROWSER:
             raise YTMusicUserError("Please provide browser authentication before using this function")
-        if not os.path.isfile(filepath):
+        fp = Path(filepath)
+        if not fp.is_file():
             raise YTMusicUserError("The provided file does not exist.")
 
         supported_filetypes = ["mp3", "m4a", "wma", "flac", "ogg"]
-        if os.path.splitext(filepath)[1][1:] not in supported_filetypes:
+        if fp.suffix[1:] not in supported_filetypes:
             raise YTMusicUserError(
                 "The provided file type is not supported by YouTube Music. Supported file types are "
                 + ", ".join(supported_filetypes)
@@ -224,8 +224,12 @@ class UploadsMixin(MixinProtocol):
 
         headers = self.headers.copy()
         upload_url = f"https://upload.youtube.com/upload/usermusic/http?authuser={headers['x-goog-authuser']}"
-        filesize = os.path.getsize(filepath)
-        body = ("filename=" + ntpath.basename(filepath)).encode("utf-8")
+        filesize = fp.stat().st_size
+        if filesize >= 314572800:  # 300MB in bytes
+            msg = f"File {fp} has size {filesize} bytes, which is larger than the limit of 300MB"
+            raise YTMusicUserError(msg)
+
+        body = ("filename=" + fp.name).encode("utf-8")
         headers.pop("content-encoding", None)
         headers["content-type"] = "application/x-www-form-urlencoded;charset=utf-8"
         headers["X-Goog-Upload-Command"] = "start"
@@ -235,7 +239,7 @@ class UploadsMixin(MixinProtocol):
         headers["X-Goog-Upload-Command"] = "upload, finalize"
         headers["X-Goog-Upload-Offset"] = "0"
         upload_url = response.headers["X-Goog-Upload-URL"]
-        with open(filepath, "rb") as file:
+        with open(fp, "rb") as file:
             response = requests.post(upload_url, data=file, headers=headers, proxies=self.proxies)
 
         if response.status_code == 200:
@@ -243,7 +247,7 @@ class UploadsMixin(MixinProtocol):
         else:
             return response
 
-    def delete_upload_entity(self, entityId: str) -> Union[str, Dict]:  # pragma: no cover
+    def delete_upload_entity(self, entityId: str) -> Union[str, dict]:  # pragma: no cover
         """
         Deletes a previously uploaded song or album